+++ src/deluge/gui/l10n/g_english.cpp
@@ -809,6 +809,7 @@ PLACE_SDRAM_DATA Language english{
         {STRING_FOR_DEFAULT_CLIP_TYPE, "Clip Type"},
         {STRING_FOR_DEFAULT_NEW_CLIP_TYPE, "New Clip Type"},
         {STRING_FOR_DEFAULT_USE_LAST_CLIP_TYPE, "Use Last Clip Type"},
+        {STRING_FOR_DEFAULT_STEREO_VU_METER, "Stereo VU Meter"},
     },
 };
 } // namespace deluge::l10n::built_in