+++ src/deluge/gui/views/arranger_view.cpp
@@ -947,7 +947,8 @@ ActionResult ArrangerView::padAction(int32_t x, int32_t y, int32_t velocity) {
 
 	// don't interact with sidebar if VU Meter is displayed
 	// and you're in the volume/pan mod knob mode (0)
-	if (x >= kDisplayWidth && view.displayVUMeter && (view.getModKnobMode() == 0)) {
+	bool xDisplayForVUMeter = FlashStorage::defaultStereoVUMeter ? kDisplayWidth : kDisplayWidth + 1;
+	if (x >= xDisplayForVUMeter && view.displayVUMeter && (view.getModKnobMode() == 0)) {
 		return ActionResult::DEALT_WITH;
 	}