+++ src/deluge/gui/views/performance_session_view.cpp
@@ -1000,9 +1000,11 @@ ActionResult PerformanceSessionView::padAction(int32_t xDisplay, int32_t yDispla
 		else if ((xDisplay >= kDisplayWidth) && !defaultEditingMode) {
 			// don't interact with sidebar if VU Meter is displayed
 			// and you're in the volume/pan mod knob mode (0)
-			if (view.displayVUMeter && (view.getModKnobMode() == 0)) {
+			bool xDisplayForVUMeter = FlashStorage::defaultStereoVUMeter ? kDisplayWidth : kDisplayWidth + 1;
+			if (xDisplay >= xDisplayForVUMeter && view.displayVUMeter && (view.getModKnobMode() == 0)) {
 				return ActionResult::DEALT_WITH;
 			}
+
 			// if in arranger view
 			if (currentSong->lastClipInstanceEnteredStartPos != -1) {
 				// pressing the first column in sidebar to trigger sections / clips