+++ src/deluge/storage/flash_storage.cpp
@@ -242,6 +242,8 @@ bool accessibilityMenuHighlighting = true;
 OutputType defaultNewClipType = OutputType::SYNTH;
 bool defaultUseLastClipType = true;
 
+bool defaultStereoVUMeter = true;
+
 void resetSettings() {
 
 	cvEngine.setCVVoltsPerOctave(0, 100);
@@ -342,6 +344,8 @@ void resetSettings() {
 
 	defaultNewClipType = OutputType::SYNTH;
 	defaultUseLastClipType = true;
+
+	defaultStereoVUMeter = true;
 }
 
 void resetMidiFollowSettings() {
@@ -747,6 +751,13 @@ void readSettings() {
 	else {
 		defaultUseLastClipType = buffer[177];
 	}
+
+	if (buffer[178] != 0 && buffer[178] != 1) {
+		defaultStereoVUMeter = true;
+	}
+	else {
+		defaultStereoVUMeter = buffer[178];
+	}
 }
 
 static bool areMidiFollowSettingsValid(std::span<uint8_t> buffer) {
@@ -1015,6 +1026,8 @@ void writeSettings() {
 
 	buffer[176] = util::to_underlying(defaultNewClipType);
 	buffer[177] = defaultUseLastClipType;
+
+	buffer[178] = defaultStereoVUMeter;
 
 	R_SFLASH_EraseSector(0x80000 - 0x1000, SPIBSC_CH, SPIBSC_CMNCR_BSZ_SINGLE, 1, SPIBSC_OUTPUT_ADDR_24);
 	R_SFLASH_ByteProgram(0x80000 - 0x1000, buffer.data(), 256, SPIBSC_CH, SPIBSC_CMNCR_BSZ_SINGLE, SPIBSC_1BIT,